Cleanup TrimmingAVPlayerView
